# ff-split/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinFlow Split Service API
  description: |
    API для сервиса управления совместными расходами в FinFlow.
    
    Сервис предоставляет функциональность для:
    - Управления мероприятиями и событиями
    - Создания и отслеживания транзакций
    - Разделения расходов между пользователями
    - Управления задачами и активностями
    - Работы с категориями и иконками
    
    ## Аутентификация
    
    Защищенные endpoints требуют аутентификации через Bearer токен:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: FinFlow Team
  license:
    name: MIT
servers:
  - url: http://localhost:8081
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/event/{id_event}/user:
    get:
      tags:
        - users
      summary: Получить пользователей мероприятия
      description: Возвращает список пользователей мероприятия
      operationId: getUsersByEventID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - users
      summary: Добавить пользователей в мероприятие
      description: Добавляет пользователей в мероприятие
      operationId: addUsersToEvent
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUsersRequest'
      responses:
        '200':
          description: Пользователи добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/user/{id_user}:
    delete:
      tags:
        - users
      summary: Удалить пользователя из мероприятия
      description: Удаляет пользователя из мероприятия
      operationId: removeUserFromEvent
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: id_user
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользователь удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/user/external:
    get:
      tags:
        - users
      summary: Получить внутренние ID пользователей по внешним ID
      description: Возвращает внутренние ID пользователей по списку внешних ID. Если пользователь не найден в системе, синхронизирует его с ff-id сервисом
      operationId: getUsersByExternalIDs
      parameters:
        - name: uids
          in: query
          required: true
          schema:
            type: array
            items:
              type: integer
              format: int64
          style: form
          explode: false
          description: Список внешних ID пользователей
      responses:
        '200':
          description: Список пользователей с внутренними и внешними ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/internal/{id_user}:
    get:
      tags:
        - users
      summary: Получить профиль пользователя
      description: Возвращает информацию о профиле пользователя
      operationId: getUserByID
      parameters:
        - name: id_user
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/sync:
    post:
      tags:
        - users
      summary: Синхронизировать пользователей
      description: Синхронизирует информацию о пользователях с ff-id сервисом
      operationId: syncUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncUsersRequest'
      responses:
        '200':
          description: Пользователи синхронизированы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/user/dummy:
    post:
      tags:
        - users
      summary: Создать dummy-пользователя
      description: Создает dummy-пользователя в мероприятии
      operationId: createDummyUser
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DummyUserRequest'
      responses:
        '201':
          description: Dummy-пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event/{id_event}/user/dummies:
    get:
      tags:
        - users
      summary: Получить dummy-пользователей
      description: Возвращает список dummy-пользователей мероприятия
      operationId: getDummiesByEventID
      parameters:
        - name: id_event
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список dummy-пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Авторизация через Bearer токен
  schemas:
    ErrorResponse:
      type: object
      required:
        - id
        - error
      properties:
        id:
          type: string
          description: ID запроса
        error:
          $ref: '#/components/schemas/ErrorResponseDetail'
    ErrorResponseDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
          example: "validation"
        message:
          type: string
          description: Сообщение об ошибке
          example: "invalid input"
        data:
          $ref: '#/components/schemas/ErrorResponseAdditionalData'
    ErrorResponseAdditionalData:
      type: object
      properties:
        slug:
          type: string
          description: Дополнительная информация
          example: "slug"
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Статус успеха операции
...
    UserProfileDTO:
      type: object
      properties:
        internal_id:
          type: integer
          format: int64
          description: Внутренний ID пользователя в системе
        user_id:
          type: integer
          format: int64
          description: Внутренний ID пользователя
        nickname:
          type: string
          description: Никнейм пользователя
        name:
          type: string
          description: Имя пользователя
        photo:
          type: string
          description: UUID фото пользователя
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserProfileDTO'
    AddUsersRequest:
      type: object
      required:
        - user_ids
      properties:
        user_ids:
          type: array
          items:
            type: integer
            format: int64
          description: Список ID пользователей для добавления
    SyncUsersRequest:
      type: object
      required:
        - user_ids
      properties:
        user_ids:
          type: array
          items:
            type: integer
            format: int64
          description: Список ID пользователей для синхронизации
    DummyUserRequest:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
          description: Никнейм dummy-пользователя
...